Rename `types.svelte.ts` to the clearer `svelteflow-types.ts`
