+++ ts-apps/decision-logic-visualizer/src/lib/displayers/flow/layout.ts
@@ -1,7 +1,7 @@
 import dagre from '@dagrejs/dagre'
 import type { Edge } from '@xyflow/svelte'
 import { Position } from '@xyflow/svelte'
-import { isSFGroupingNode, type LadderSFNodeWithDims } from './types.svelte.js'
+import { isSFGroupingNode, type LadderSFNodeWithDims } from './svelteflow-types.js'
 
 export interface DagreConfig {
   // TODO: Clean this up in the future!
