+++ ts-apps/decision-logic-visualizer/src/lib/displayers/flow/sf-custom-nodes/bundling-source.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { Handle } from '@xyflow/svelte'
   import type { NodeProps, Node } from '@xyflow/svelte'
-  import { defaultSFHandlesInfo } from '../types.svelte.js'
+  import { defaultSFHandlesInfo } from '../svelteflow-types.js'
   // eslint-disable-next-line no-empty-pattern
   let {}: NodeProps<Node> = $props()
 </script>
